renamed package share to shared, moved method "printInteger" to package shared
