+++ src/main/scala/dotty/core/ImplicitsDemo.scala
@@ -2,9 +2,7 @@ package dotty.core
 
 import implicits.given 
 import scala.language.implicitConversions
-
-def printInteger(i:java.lang.Integer):Unit = 
-  println(s"I am an java.lang.Integer: ${i.toString}")
+import shared.printInteger
 
 @main def implicitsDemo : Unit =
   val i = 1
