+++ src/main/scala/scala2/core/MonadDemo.scala
@@ -3,7 +3,7 @@ package scala2.core
 import scala2.core.monad.MonadInstances._
 import scala2.core.monad.MonadInterfaceSyntax._  //with this import we will import everything..including the implicits
 import scala2.core.monad.Monad
-import share.MyClass
+import shared.MyClass
 
 object MonadDemo extends App {
