+++ src/main/scala/shared/CustomFunctions.scala
@@ -1,0 +1,4 @@
+package shared
+
+def printInteger(i:java.lang.Integer):Unit = 
+  println(s"I am an java.lang.Integer: ${i.toString}")
