+++ src/main/scala/shared/CustomTypeConstructors.scala
@@ -1,4 +1,4 @@
-package share
+package shared
 
 class MyClass[A](val value : A){
   override def toString = s"MyClass with type ${value.getClass} and value $value"
